+++ src/PIL/GifImagePlugin.py
@@ -25,6 +25,8 @@
 #
 
 import itertools
+import os
+import subprocess
 
 from . import Image, ImageChops, ImageFile, ImagePalette, ImageSequence
 from ._binary import i8, i16le as i16, o8, o16le as o16
@@ -265,6 +267,7 @@ class GifImageFile(ImageFile.ImageFile):
                 self.dispose = None
             elif self.disposal_method == 2:
                 # replace with background colour
+                Image._decompression_bomb_check(self.size)
                 self.dispose = Image.core.fill("P", self.size, self.info["background"])
             else:
                 # replace with previous contents
@@ -616,24 +619,22 @@ def _save_netpbm(im, fp, filename):
     # If you need real GIF compression and/or RGB quantization, you
     # can use the external NETPBM/PBMPLUS utilities.  See comments
     # below for information on how to enable this.
-
-    import os
-    from subprocess import Popen, check_call, PIPE, CalledProcessError
-
     tempfile = im._dump()
 
     with open(filename, "wb") as f:
         if im.mode != "RGB":
             with open(os.devnull, "wb") as devnull:
-                check_call(["ppmtogif", tempfile], stdout=f, stderr=devnull)
+                subprocess.check_call(["ppmtogif", tempfile], stdout=f, stderr=devnull)
         else:
             # Pipe ppmquant output into ppmtogif
             # "ppmquant 256 %s | ppmtogif > %s" % (tempfile, filename)
             quant_cmd = ["ppmquant", "256", tempfile]
             togif_cmd = ["ppmtogif"]
             with open(os.devnull, "wb") as devnull:
-                quant_proc = Popen(quant_cmd, stdout=PIPE, stderr=devnull)
-                togif_proc = Popen(
+                quant_proc = subprocess.Popen(
+                    quant_cmd, stdout=subprocess.PIPE, stderr=devnull
+                )
+                togif_proc = subprocess.Popen(
                     togif_cmd, stdin=quant_proc.stdout, stdout=f, stderr=devnull
                 )
 
@@ -642,11 +643,11 @@ def _save_netpbm(im, fp, filename):
 
             retcode = quant_proc.wait()
             if retcode:
-                raise CalledProcessError(retcode, quant_cmd)
+                raise subprocess.CalledProcessError(retcode, quant_cmd)
 
             retcode = togif_proc.wait()
             if retcode:
-                raise CalledProcessError(retcode, togif_cmd)
+                raise subprocess.CalledProcessError(retcode, togif_cmd)
 
     try:
         os.unlink(tempfile)