+++ src/PIL/TiffImagePlugin.py
@@ -854,7 +854,7 @@ class ImageFileDirectory_v2(MutableMapping):
 
         # pass 2: write entries to file
         for tag, typ, count, value, data in entries:
-            if DEBUG > 1:
+            if DEBUG:
                 print(tag, typ, count, repr(value), repr(data))
             result += self._pack("HHL4s", tag, typ, count, value)
 
@@ -1079,19 +1079,6 @@ class TiffImageFile(ImageFile.ImageFile):
         """Return the current frame number"""
         return self.__frame
 
-    @property
-    def size(self):
-        return self._size
-
-    @size.setter
-    def size(self, value):
-        warnings.warn(
-            "Setting the size of a TIFF image directly is deprecated, and will"
-            " be removed in a future version. Use the resize method instead.",
-            DeprecationWarning,
-        )
-        self._size = value
-
     def load(self):
         if self.use_load_libtiff:
             return self._load_libtiff()
@@ -1239,8 +1226,8 @@ class TiffImageFile(ImageFile.ImageFile):
             print("- YCbCr subsampling:", self.tag.get(530))
 
         # size
-        xsize = self.tag_v2.get(IMAGEWIDTH)
-        ysize = self.tag_v2.get(IMAGELENGTH)
+        xsize = int(self.tag_v2.get(IMAGEWIDTH))
+        ysize = int(self.tag_v2.get(IMAGELENGTH))
         self._size = xsize, ysize
 
         if DEBUG: