+++ src/display.c
@@ -319,18 +319,23 @@ PyImaging_DisplayModeWin32(PyObject* self, PyObject* args)
 /* -------------------------------------------------------------------- */
 /* Windows screen grabber */
 
+typedef HANDLE(__stdcall* Func_SetThreadDpiAwarenessContext)(HANDLE);
+
 PyObject*
 PyImaging_GrabScreenWin32(PyObject* self, PyObject* args)
 {
-    int width, height;
-    int includeLayeredWindows = 0;
+    int x = 0, y = 0, width, height;
+    int includeLayeredWindows = 0, all_screens = 0;
     HBITMAP bitmap;
     BITMAPCOREHEADER core;
     HDC screen, screen_copy;
     DWORD rop;
     PyObject* buffer;
+    HANDLE dpiAwareness;
+    HMODULE user32;
+    Func_SetThreadDpiAwarenessContext SetThreadDpiAwarenessContext_function;
 
-    if (!PyArg_ParseTuple(args, "|i", &includeLayeredWindows))
+    if (!PyArg_ParseTuple(args, "|ii", &includeLayeredWindows, &all_screens))
         return NULL;
 
     /* step 1: create a memory DC large enough to hold the
@@ -339,8 +344,32 @@ PyImaging_GrabScreenWin32(PyObject* self, PyObject* args)
     screen = CreateDC("DISPLAY", NULL, NULL, NULL);
     screen_copy = CreateCompatibleDC(screen);
 
-    width = GetDeviceCaps(screen, HORZRES);
-    height = GetDeviceCaps(screen, VERTRES);
+    // added in Windows 10 (1607)
+    // loaded dynamically to avoid link errors
+    user32 = LoadLibraryA("User32.dll");
+    SetThreadDpiAwarenessContext_function =
+            (Func_SetThreadDpiAwarenessContext)
+            GetProcAddress(user32, "SetThreadDpiAwarenessContext");
+    if (SetThreadDpiAwarenessContext_function != NULL) {
+        // DPI_AWARENESS_CONTEXT_PER_MONITOR_AWARE = ((DPI_CONTEXT_HANDLE)-3)
+        dpiAwareness = SetThreadDpiAwarenessContext_function((HANDLE) -3);
+    }
+
+    if (all_screens) {
+        x = GetSystemMetrics(SM_XVIRTUALSCREEN);
+        y = GetSystemMetrics(SM_YVIRTUALSCREEN);
+        width = GetSystemMetrics(SM_CXVIRTUALSCREEN);
+        height = GetSystemMetrics(SM_CYVIRTUALSCREEN);
+    } else {
+        width = GetDeviceCaps(screen, HORZRES);
+        height = GetDeviceCaps(screen, VERTRES);
+    }
+
+    if (SetThreadDpiAwarenessContext_function != NULL) {
+        SetThreadDpiAwarenessContext_function(dpiAwareness);
+    }
+
+    FreeLibrary(user32);
 
     bitmap = CreateCompatibleBitmap(screen, width, height);
     if (!bitmap)
@@ -354,7 +383,7 @@ PyImaging_GrabScreenWin32(PyObject* self, PyObject* args)
     rop = SRCCOPY;
     if (includeLayeredWindows)
         rop |= CAPTUREBLT;
-    if (!BitBlt(screen_copy, 0, 0, width, height, screen, 0, 0, rop))
+    if (!BitBlt(screen_copy, 0, 0, width, height, screen, x, y, rop))
         goto error;
 
     /* step 3: extract bits from bitmap */
@@ -376,7 +405,7 @@ PyImaging_GrabScreenWin32(PyObject* self, PyObject* args)
     DeleteDC(screen_copy);
     DeleteDC(screen);
 
-    return Py_BuildValue("(ii)N", width, height, buffer);
+    return Py_BuildValue("(ii)(ii)N", x, y, width, height, buffer);
 
 error:
     PyErr_SetString(PyExc_IOError, "screen grab failed");